Update layout.js
Today, I reset the wordpress environment because of the only option remained After following other options did not help me to get the issue resolved for resolving the error which I was facing for two days. Now, I am waiting for the Ma'am Aisha to install the Pro version of ACF, on the base of which this project architecture stands, which only Ma'am has the authority to install.